Add files via upload
For this study I use the Data from the last 20 years of the index S&P500 and MSCI Wolrd. 
For the Insurance I use the standard return of a policy of 20 years.

A better study can comprehend the data of inflation for a better comperrison of the value of the money today and at the end of the investing.